fix failing test on osx
git-svn-id: https://svn.java.net/svn/jna~svn/trunk@930 2f8a963e-d2e4-e7d0-97bf-ccb7fcea9d80

+++ jnalib/test/com/sun/jna/RawCallbacksTest.java
@@ -56,7 +56,7 @@ public class RawCallbacksTest extends CallbacksTest {
     }
     
     public static class DirectCallbackTestLibrary implements CallbackTestLibrary {
-        public native float callInt32Callback(Int32Callback c, float arg, float arg2);
+        public native double callInt32Callback(Int32Callback c, double arg, double arg2);
         static {
             Native.register(NativeLibrary.getInstance("testlib", _OPTIONS));
         }
